requirements.txt
python-fasthtml==0.4.4
sqlite_minutils==4.0.3
uvicorn==0.30.1
yt-dlp==2025.10.22
moviepy==2.2.1

